:construction_worker: upgrade stack Spring Cloub Finchley.SR2
