+++ .springBeans
@@ -7,9 +7,9 @@
 	</configSuffixes>
 	<enableImports><![CDATA[false]]></enableImports>
 	<configs>
-		<config>java:com.scc.zuulsvr.ZuulServerApplication</config>
 	</configs>
 	<autoconfigs>
+		<config>java:com.scc.zuulsvr.controllers.DocumentationController</config>
 	</autoconfigs>
 	<configSets>
 	</configSets>
