+++ README.md
@@ -1,0 +1,7 @@
+## Spring Boot
+
+2.0.1.RELEASE
+
+## Spring Cloud Stack
+
+version [`Finchley.SR2`](https://spring.io/projects/spring-cloud-netflix)
