+++ src/main/docker/run.sh
@@ -7,23 +7,9 @@ while ! `nc -z eurekaserver  $EUREKASERVER_PORT`; do sleep 3; done
 echo "******* Eureka Server has started"
 
 echo "********************************************************"
-echo "Waiting for the configuration server to start on port $CONFIGSERVER_PORT"
-echo "********************************************************"
-while ! `nc -z configserver $CONFIGSERVER_PORT`; do sleep 3; done
-echo "*******  Configuration Server has started"
-
-echo "********************************************************"
-echo "Waiting for the ZIPKIN server to start  on port $ZIPKIN_PORT"
-echo "********************************************************"
-while ! `nc -z zipkin $ZIPKIN_PORT`; do sleep 10; done
-echo "******* ZIPKIN has started"
-
-echo "********************************************************"
-echo "Starting Zuul Service with $CONFIGSERVER_URI"
+echo "Starting Zuul Service
 echo "********************************************************"
 java -Djava.security.egd=file:/dev/./urandom -Dserver.port=$SERVER_PORT   \
      -Deureka.client.serviceUrl.defaultZone=$EUREKASERVER_URI   \
-     -Dspring.cloud.config.uri=$CONFIGSERVER_URI                \
      -Dspring.profiles.active=$PROFILE                          \
-     -Dspring.zipkin.baseUrl=$ZIPKIN_URI                        \
      -jar /usr/local/zuulservice/@project.build.finalName@.jar
