+++ src/main/java/com/scc/zuulsvr/config/ServiceConfig.java
@@ -1,11 +1,9 @@
 package com.scc.zuulsvr.config;
 
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cloud.context.config.annotation.RefreshScope;
 import org.springframework.stereotype.Component;
 
 @Component
-@RefreshScope
 public class ServiceConfig{
 
   @Value("${authentification.max.attempt}")
