+++ src/test/java/com/scc/zuulsvr/ApplicationTest.java
@@ -6,8 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
-@SpringBootTest(classes = ZuulServerApplication.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class ZuulServerApplicationTest {
+@SpringBootTest(classes = Application.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class ApplicationTest {
 
 	/*
 	 * Référence documentation: http://www.baeldung.com/spring-boot-testing
